+++ engine/src/main/java/dev/gamekit/core/Component.java
@@ -40,8 +40,8 @@ public abstract class Component {
    */
   protected void start() { /* No-op */ }
 
-  /** Called when the component is restarted after previously being stopped */
-  protected void restart() { /* No-op */ }
+  /** Called when the component is resumed after previously being stopped */
+  protected void resume() { /* No-op */ }
 
   /** Called to update the component */
   protected void update() { /* No-op */ }
@@ -52,7 +52,7 @@ public abstract class Component {
   /**
    * Called to stop the component, but not dispose it.
    * <p>
-   * The component can be restarted at a later time via {@link #restart}
+   * The component can be resumed at a later time via {@link #resume}
    */
   protected void stop() { /* No-op */ }
 
@@ -79,9 +79,9 @@ public abstract class Component {
     start();
   }
 
-  /** Called <b>once</b> by the host {@link Entity} to restart the component */
-  void _restart() {
-    restart();
+  /** Called by the host {@link Entity} to resume the component after being previously stopped */
+  void _resume() {
+    resume();
   }
 
   /** Called by the host {@link Entity} to update the component */